Fixes issues with migrating to app router functionality
