+++ app/index.tsx
@@ -1,12 +1,29 @@
-import { View, Text } from "react-native";
-import React from "react";
+import React, { useLayoutEffect } from "react";
+import { SafeAreaView, ScrollView } from "react-native";
+import { useNavigation } from "@react-navigation/native";
 
-const Page = () => {
+import tw from "twrnc";
+
+import Header from "../components/Header";
+
+const Index = () => {
+  const navigation = useNavigation();
+
+  useLayoutEffect(() => {
+    navigation.setOptions({
+      headerShown: false,
+    });
+  }, []);
+
   return (
-    <View>
-      <Text>index</Text>
-    </View>
+    <SafeAreaView style={tw`bg-white pt-5`}>
+      <Header />
+      <ScrollView
+        style={tw`bg-gray-100`}
+        contentContainerStyle={{ paddingBottom: 100 }}
+      ></ScrollView>
+    </SafeAreaView>
   );
 };
 
-export default Page;
+export default Index;
