+++ babel.config.js
@@ -1,10 +1,11 @@
 module.exports = function (api) {
   api.cache(true);
   return {
-    presets: ['babel-preset-expo'],
+    presets: ["babel-preset-expo"],
     plugins: [
       // Required for expo-router
-      'expo-router/babel',
+      "expo-router/babel",
+      "react-native-reanimated/plugin",
     ],
   };
 };
